Fix allowCond to control the 3 possible options
When decoding branches, it's useful to control if you want conditional
branches, unconditional branches, or any of them. By only using 1
boolean it was tricky to know if it was a conditional branch.

This way it's more explicit.

Also solve the isCall issue where it was defining if LR was being used
wrongly on relocation.