Use lwz when an immediate fits 32 bits
In order to avoid waste of memory in TOC, the minimum element size
in TOC is now 32 bits. If an immediate fits in 32 bit a lwz is
emmitted. If does not fit, a ld is emmitted.

If one does not want to use this new feature, he/she can use the
option -vEval.ppc64useTOCLwz=1/0 .