Remove NO_OPT workaround from decoder
Summary:

This commit removes __attribute__(optimize("O0")) of decoder and
remove early-inlining-insns compilation options for ppc64.

Compilation freezes due a fine tune optimization
using early-inlining-insns and -O3 optimization level on POWER.
Before we provide a temporary solution disabling optimization for
decoder constructor (who have lots of inline and macro stuffs).

After some investigation and interlock with gcc developers we found
that's caused by early-inlining-insns due a gcc bug on ppc64el.
A bugzilla was openned on:

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=72855

Was pointed out we should be careful about those parameters as it
may have to be fine tunned for each architecture or we can find
problems like that.