Fix index
Then the index is bigger than UINT16_MAX, the instruction addis must be
used in order to sum part of the offset. If the last four bytes is still
bigger than INT16_MAX, it means that the value is negative and it is its
complement, so the number being added in the addis instruction must be
added by one.