Correctly store the source address of decoded instruction
On disassembly, it may happen that the pointer to a instruction is not
actually the emitted instruction, rather only a buffer read with that
value (e.g: tc-print behavior). For that reason, another parameter needs
to be passed in order to relate to the correct addressing, or else
interpretation of relative jumps is compromised (e.g).