Fix backchain for PPC64
Entering in the TC area was breaking the backchain creating an infinite
loop. The issue was that the frame created by the enterTCImpl was not
being closed in the emitEnterTCHelper and a new was not being created.
The fix changes the inittc to close the frame and create a new one to be
used in the TC area. Consequently the inittc position had to be changed
to be the first vasm emitted in the emitEnterTCHelper.