Enable REMOVE_FAR_BRANCHES_NOPS in toc+relocation
In order to enable this option, it is necessary to set the "call"
on "calls" as smashable.

Another issue that had to be solved was to keep a "patchable" size
on limmediate when a code is relocated and TOC gets unexpectedly
full. In this case, an extra nop has to be emited.