Change limmediate to have a state machine as param
This comment describes it neatly:
enum class ImmType {
              // Supports TOC? | Supports li64? | Fixed size?
              // -------------------------------------------
  AnyCompact, // Yes           | Yes            | No
  AnyFixed,   // Yes           | Yes            | Yes (5 instr)
  TocOnly,    // Yes           | No             | Yes (2 instr)
};

Therefore, when immediate is called, this parameter has to be set. It'll
define if it's as small as possible (AnyCompact), or as it was before
for fixedSize (AnyFixed) or always using TOC and reserving 2
instructions for it (TocOnly).

Smashable instructions will always use TocOnly.

Without relocation, no quick test regression.
With relocation, about 180 of 1099 tests are failing. :(