Remove strcmp-power8.S
Summary:

Turns out LGPL license is not compatible to PHP license:

https://www.gnu.org/licenses/license-list.en.html

"PHP License, Version 3.01 (#PHP-3.01)

This license is used by most of PHP4. It is a non-copyleft
free software license. It is incompatible with the GNU GPL
because it includes strong restrictions on the use of “PHP”
in the name of derived products.
We recommend that you not use this license for anything
except PHP add-ons."

The pull request for this was rejected by FB due this.

I guess this makes more sense when HHVM only compiles with
old glibc. I guess new versions of glibc will use this
version of strcmp anyway on POWER.