Fix #115 - Fix GP args to include r10
Somehow only r3-r9 were defined although the ABI points out that r10 is
also a GP arg. This didn't bring us to problems before as it's rare to
call a function with so many arguments however the msg_receive function
is one of them and the garbage was invalid when running hhvm from
/usr/local/bin/hhvm.

When calling a function with 8 GP args, HHVM was spilling the 8th
argument to the stack (and keeping 16-bytes alignment, of course) before
setting all the other register arguments:
addi    r3,r31,-128
addi    r27,r27,-8
stdu    r3,-8(r27)

But msg_receive was reading from r10 even if it was undefined, therefore
garbage was being used. Now it correctly sets it:
addi    r10,r31,-128