Fix zero-extension vasms on PPC64.
Summary:

Vasms that should zero-extend small registers to bigger registers was using
sign extending instruction (extrb) which is not quit right.
We try to change to use mov instructions instead but when map movzbl and movzbq
to movb we got some fail tests on quick suite in Debug mode.

The problem was the instruction `movb` was being removed as a `dead code`
since it was considered a trivial nop if `movb r(dst), r(src) with dst = src`.

On PPC64 the `movb` vasm instruction is mapped to a `rlwinm` instruction that
also zero upper bits of register and is used in `movzb` type instructions
unlike x86_64. Hence we cannot consider `movb` as a trivial nop on PPC64 even
if r(dst) = r(src) since we may alter the register content.