Fix (signal) extensions and move vasms.
Summary:

This commit fixes move and (signal) extensions vasms. Some vasms was being used
to move and extend (signal) which can introduce wrong behaviours. In many cases
the signal extension was useless (only extends '0').
Now we use signal extensions only where it's needed so we emit one instruction
less when we need to move but not sign ext.

This commit also fixes:

- Vasm extrw was implemented as movw and do not sign extended. When change to
  use ppc64 instruction extsw got a invalid convertion Reg64 to Reg16.
- movzwq instruction was lowered to extrw -- using movw instruction instead.
- Remove some wierd calculations for mask size when map mov* instructions to
  ppc64 asm.