Fix sign ext/mov vasms for PPC64.
Summary:

This commit fixes move and (signal) extensions vasms for PPC64. Some vasms
that needs to move values less than 64 bits to 64 bits registers.

Those vasms were tranlated to pairs that mov and sign extend (rlwinm + exts*)
that are used even when a sign extended are not intended like movz* or
bitwise (xor*, and*) which can lead to a wrong behavior or emit a extra useless
sign extend instruction.

This commit fixes that problem, now we only extend sign when it's needed.
To summarize this fix:

- Do not consider a movb with r(src) = r(dst) as a trivial nop on PPC64 as
movb is mapped to a ppc64 instruction (rlwinm) that always zeros upper bits and
is used in different places unlike x64.
- Vasm extrw was implemented by mwillians as a movw that do not sign extend and
used on movzwq instructions. Now movzwq uses a movw instruction and extrw was
changed to do movw and sign extend.
- When change extrw to use ppc64 instruction extsw got a invalid cast Reg64 to
Reg16. This was fixed adding a extra (implicit) constructor on
vasm-reg.h.
- Remove some wierd calculations for mask size when map mov* instructions to
  ppc64 asm.
- Fix extsw, extsh to include rc param even if we don't emit rc (CR record)
version of those instructions.