Improve thread safety of PPC64's decoder
If 2 threads are decoding the same kind of instruction, a race condition
may happen and the DecoderInfo returned would be the same for both,
despite being possible to have different parameters.

This change set the parameters for each call on different local
instances. It may also improve performance as it's not writing back to
the global vector with the newest parameters, which is actually not
necessary at all.