Avoid non-volatile r30 register as it's not saved
Starting on gcc v5.4.1, the r30 register can't be clobbered and the ABI
still flags it as a non-volatile register.

Recently some tests started to fail as the value of r30 could not be
lost when executing the VM. For some reason the compiler was not
generating a code which r30 was being used so this fault was not
happening.