Enable TOC to any immediate
Using the instructions li/lis to load an immediate causes race condition
when smashing it, as the immediate is splited into up to 5 load
instructions, then it makes read/write not atomic.
Using the TOC (Table Of Contents) concept, which is a memory space with the
base address stored in the R2 register, the immediate is stored in one
memory position, making the read/write atomic operations, as only one
instruction is necessary to read or modify it.
This first change makes it work, removing the optimizations, as they are
not compatible with the smash instructions.