Avoid insert duplicate values in TOC
Many instructions uses the same immediate value, as -1, 0, 1, etc., so in
order to avoid insertion of duplicated values, a map was created to verify
already inserted values, and use them.
The particular case of smash instructions is being treated with new
insertion, i.e., does not use the same immediate that would have being
inserted, as it may change after relocate or patch it.